+++ CHANGELOG.md
@@ -55,6 +55,7 @@
    * ADDED: mostly global graph attributes to mjolnir's statsd logging [#6021](https://github.com/valhalla/valhalla/pull/6021)
    * ADDED: free flow and constrained flow speeds to mvt edge layer [#6014](https://github.com/valhalla/valhalla/pull/6014)
    * CHANGED: perf(costmatrix): use `ankerl::unordered_dense::set` instead of `std::set` for `LocationStatus::unfound_connections` [#6013](https://github.com/valhalla/valhalla/pull/6013)
+   * ADDED: `openlr` module to Python bindings [#6087](https://github.com/valhalla/valhalla/pull/6087)
 
 ## Release Date: 2026-02-19 Valhalla 3.6.3
 * **Removed**